src/main/java/io/zipcoder/casino/utilities/CasinoGame/BlackJackPlayer.java
package io.zipcoder.casino.utilities.CasinoGame;
//import java.io.Console;
import java.io.InputStream;
import java.io.PrintStream;
public class BlackJackPlayer extends Player implements GamblingPlayer {

    Console console = new Console(System.in,System.out);
    String bjname;
    Integer bjbalance;

    public  BlackJackPlayer(){

    }
    public  BlackJackPlayer( Player player){

    }

  public BlackJackPlayer(String name ,Integer balance){

      super( name ,  balance );

      bjname = name;

       bjbalance = balance;
}

    public void placeBet(Integer betValue) {

        Integer remainingbalance = balance- betValue;

    }


    @Override
    public void receiveWinnings() {

    }
    public Integer recieveWinnings() {

        Integer integerInput = console.getIntegerInput("");

            return integerInput;
        }


}




